+++ app/app.go
@@ -28,11 +28,11 @@ import (
 const (
 	appName                = "IrisApp"
 	appPrometheusNamespace = "iris"
-	// FlagReplay used for replaying last block flag
+	// FlagReplay used for replaying last block
 	FlagReplay = "replay-last-block"
-	// DefaultSyncableHeight Multistore saves a snapshot every 10000 blocks
+	// Multistore saves a snapshot every 10000 blocks
 	DefaultSyncableHeight = store.NumStoreEvery
-	// DefaultCacheSize Multistore saves last 100 blocks
+	// Multistore saves last 100 blocks
 	DefaultCacheSize = store.NumRecent
 )
 
@@ -66,13 +66,6 @@ func NewIrisApp(logger log.Logger, db dbm.DB, config *cfg.InstrumentationConfig,
 	if viper.GetBool(FlagReplay) {
 		lastHeight := Replay(app.Logger)
 		err = app.LoadVersion(lastHeight, protocol.KeyMain, true)
-
-		// If reset to another protocol version, should reload Protocol and reset txDecoder
-		loaded, current := app.Engine.LoadCurrentProtocol(app.GetKVStore(protocol.KeyMain))
-		if !loaded {
-			cmn.Exit(fmt.Sprintf("Your software doesn't support the required protocol (version %d)!", current))
-		}
-		app.BaseApp.txDecoder = auth.DefaultTxDecoder(app.Engine.GetCurrentProtocol().GetCodec())
 	} else {
 		err = app.LoadLatestVersion(protocol.KeyMain)
 	} // app is now sealed